Revert "Add rubocop-md"
This reverts commit a30ef798c9d42750fad9ed5958a850d5382e015e.

`rubocop-md` is a unique gem that doesn't possess any cops. It inspects Ruby code within Markdown files.
However, as can be seen from the disabling of cops in `rails/rails`, which cops to disable is documented
in the inspection target repository:

```diff
  Exclude:
+   - '**/CHANGELOG.md'
+   - '**/2_*_release_notes.md'
+   - '**/3_*_release_notes.md'
+   - '**/4_*_release_notes.md'
+   - '**/5_*_release_notes.md'
+   - '**/6_*_release_notes.md'

 Layout/IndentationConsistency:
   Enabled: true
   EnforcedStyle: indented_internal_methods
+  Exclude:
+    - '**/*.md'

 Style/FrozenStringLiteralComment:
   Enabled: true
   EnforcedStyle: always
   Exclude:
     - 'actionview/test/**/*.builder'
     - 'actionview/test/**/*.ruby'
     - 'actionpack/test/**/*.builder'
     - 'actionpack/test/**/*.ruby'
     - 'activestorage/db/migrate/**/*.rb'
     - 'activestorage/db/update_migrate/**/*.rb'
     - 'actionmailbox/db/migrate/**/*.rb'
     - 'actiontext/db/migrate/**/*.rb'
+    - '**/*.md'
```

This means that any false positives in documentation need to be addressed as they arise, leading one to
question whether this is the behavior users expect.

This PR resolves the above-mentioned usability issues by disabling `rubocop-md`. If users wish to enable
`rubocop-md`, they simply need to add it explicitly to their `.rubocop.yml`. In other words, the use of
`rubocop-md` is a deliberate choice made by the user.

I believe this is a better configuration for `rubocop-rails_config`.